hy-paddle-b0-0604new/requirements.txt
albumentations>=1.1.0

matplotlib>=3.2.2
numpy>=1.21.2
opencv_python_headless>=4.5.5.62
prefetch_generator>=1.0.1
PyYAML>=3.13
scipy>=1.4.1
seaborn>=0.11.2
tensorboardX>=2.4.1

tqdm>=4.61.2
webcolors>=1.11
paddleseg
paddlepaddle